+++ Models/Sale.cs
@@ -4,7 +4,7 @@ public class Sale
 {
     public int Id { get; set; }
     public DateTime SaleDate { get; set; } = DateTime.UtcNow;
-    public decimal TotalAmount { get; set; }
+    public int TotalAmount { get; set; }
 
     public List<SaleItem> Items { get; set; } = new();
 }
